Add new "launcher" screen to serve as a common starting point for all communications methods.
The end result of a *whole* lot of stuffing around trying to get things to work.

While a launcher screen was always intended, its addition was driven in this case by a need to make up for the flags "FLAG_ACTIVITY_TASK_ON_HOME" and "FLAG_ACTIVITY_CLEAR_TASK" not being available for Gingerbread. Note: The previous code did work for non-Gingerbread situations but was a little bit less than ideal.

Adding the launcher screen necessitated a whole lot of stuffing around to try and get all permutations of not-yet-started, already-started, already-run-once, on-network-screen and usb-auto-launch-occurred working correctly. Which they now do--*but* only on Honeycomb and above.

On Gingerbread a USB connection only works once. The second time we still incur a "device not found" error when the handshake data is sent. (But this AFAIK is not due to adding the launcher.)

I'm committing this work now since it at least preserves the existing correct behaviour on Honeycomb+ and I'll look at debugging the other issue on Gingerbread separately.